Added some new particle effects for ball bounces and the beginnings of block break effects.

+++ ESPEngine/ESPParticleColourEffector.cpp
@@ -11,13 +11,13 @@ ESPParticleColourEffector::~ESPParticleColourEffector() {
 void ESPParticleColourEffector::AffectParticleOnTick(double dT, ESPParticle* particle) {
 	// Obtain the current colour and alpha of the particle
 	Colour currColour;
-	float  currAlpha;
+	double  currAlpha;
 	particle->GetColour(currColour, currAlpha);	
 	
 	// Perform linear interpolation on the alpha of the particle
 	// based on the particle's lifetime
-	float newAlpha = static_cast<float>(this->startAlpha + particle->GetCurrentLifeElapsed() * 
-		(this->endAlpha - this->startAlpha) / particle->GetLifespanLength());
+	double newAlpha = this->startAlpha + particle->GetCurrentLifeElapsed() * 
+		(this->endAlpha - this->startAlpha) / particle->GetLifespanLength();
 
 	particle->SetColour(currColour, newAlpha);
 }
